tools/office-skills/requirements.txt
# ppt-template - Python 依赖

# DOCX 操作 (docx skill)
python-docx>=1.1.0
lxml>=5.0.0

# PDF 操作 (pdf skill)
reportlab>=4.0.0
playwright>=1.40.0

# 电子表格操作 (spreadsheets skill)
# artifact_tool 是内部库，无需安装

# PPT 操作 (slide skill)
python-pptx>=1.0.0
defusedxml>=0.7.1

# 通用依赖（图片、PDF 渲染等）
pdf2image>=1.17.0
pillow>=11.0.0
numpy>=1.24.0
